requirements.txt
accelerate==1.7.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.0
aiosignal==1.3.2
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.3.0
certifi==2025.4.26
charset-normalizer==3.4.2
comm==0.2.2
datasets==3.6.0
debugpy==1.8.14
decorator==5.2.1
dill==0.3.8
exceptiongroup==1.3.0
executing==2.2.0
filelock==3.13.1
frozenlist==1.6.0
fsspec==2024.6.1
hf-xet==1.1.2
huggingface-hub==0.32.0
idna==3.10
ipykernel==6.29.5
ipython==8.36.0
jedi==0.19.2
Jinja2==3.1.4
jupyter_client==8.6.3
jupyter_core==5.7.2
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mpmath==1.3.0
multidict==6.4.4
multiprocess==0.70.16
nest-asyncio==1.6.0
networkx==3.3
numpy==2.1.2
packaging==25.0
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.3.8
prompt_toolkit==3.0.51
propcache==0.3.1
protobuf==6.31.0
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==20.0.0
Pygments==2.19.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.4.0
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
sentencepiece==0.2.0
six==1.17.0
stack-data==0.6.3
sympy==1.13.3
tiktoken==0.9.0
tokenizers==0.21.1
torchaudio
torchvision
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
git+https://github.com/huggingface/transformers.git
triton==3.3.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
wcwidth==0.2.13
xxhash==3.5.0
yarl==1.20.0
